LayoutTests/imported/w3c/web-platform-tests/css/css-text/white-space/white-space-collapse-001-expected.txt


PASS no-break space
PASS en space
PASS em space
PASS thin space
PASS zero width space
PASS narrow no-break space
PASS ideographic space
PASS ideographic space inside ideographic text
hello   there
hello   there
hello   there
hello   there
hello   there
hello   there
hello   there
hello   there
hello ​ there
hello ​ there
hello   there
hello   there
hello   there
hello   there
缔造真正全球通行   的万维网
缔造真正全球通行   的万维网
